On AVR gcc you can use different width types for bitfields
This is not safe on arm gcc and you run into really strange errors. So force everything to be uint32_t wide to be consistent on non-avr platforms.

Also propagate debug mode throughout the entire cache hierarchy so one can enable compile time debugging of the cache and all of its components